+++ config/hard/InertiaRide/code/CGCode.java
@@ -1,0 +1,24 @@
+import java.util.*;
+import java.io.*;
+import java.math.*;
+
+class Solution {
+
+    public static void main(String args[]) {
+        Scanner in = new Scanner(System.in);
+        int inertia = in.nextInt();
+        int W = in.nextInt();
+        int H = in.nextInt();
+        if (in.hasNextLine()) {
+            in.nextLine();
+        }
+        for (int i = 0; i < H; i++) {
+            String row = in.nextLine();
+        }
+
+        // Write an answer using System.out.println()
+        // To debug: System.err.println("Debug messages...");
+
+        System.out.println("answer");
+    }
+}