+++ config/hard/InertiaRide/code/CGCode.php
@@ -1,0 +1,12 @@
+<?php
+fscanf(STDIN, "%d", $inertia);
+fscanf(STDIN, "%d %d", $W, $H);
+for ($i = 0; $i < $H; $i++)
+{
+    $row = stream_get_line(STDIN, 128 + 1, "\n");
+}
+
+// Write an answer using echo(). DON'T FORGET THE TRAILING \n
+// To debug: error_log(var_export($var, true)); (equivalent to var_dump)
+
+echo("answer\n");