+++ config/hard/InertiaRide/code/CGCode.ts
@@ -1,0 +1,12 @@
+const inertia: number = parseInt(readline());
+var inputs: string[] = readline().split(' ');
+const W: number = parseInt(inputs[0]);
+const H: number = parseInt(inputs[1]);
+for (let i = 0; i < H; i++) {
+    const row: string = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');