+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "To Sky’s Edge".
 - configuration for "The hungry duck - part 1".
 - configuration for "Dominoes solver".
+- configuration for "Inertia ride".
 
 ## [1.15.0] - 2025-05-31
 ### Added